Move TrajSeedMatchingCutsPortable to RecoEgamma/EgammaElectronProducers/plugins/alpaka.
